Updated directory structure, deploy scripts, etc.
