+++ deploy/README.md
@@ -1,0 +1,7 @@
+## Deployment Instructions
+
+Currently, WIT is deployed to AWS on an EC2 instance. 
+
+
+### Directory
+
