+++ deploy/wit_app_nginx-template.conf
@@ -1,0 +1,13 @@
+## This is an example of a configuration file used for nginx; make a copy of it and add in new values if needed.
+
+server {
+	listen	80;
+	server_name servername;
+	charset	utf-8;
+	client_max_body_size	75M;
+	
+	location / {
+		include uwsgi_params;
+		uwsgi_pass unix:/var/www/wit/wit_uwsgi.sock;
+	}
+}
