+++ deploy/wit_uwsgi.ini
@@ -1,0 +1,19 @@
+[uwsgi]
+base = /var/www/wit
+
+app = recommend_api
+module = %(app)
+
+home = /home/ubuntu/wit/venv
+pythonpath = %(base)
+
+socket = %n.sock
+
+chmod-socket = 660
+vacuum = true
+
+logto = /var/log/uwsgi/%n.log
+
+callable = app
+
+die-on-term = true
