--- job_post.txt
@@ -1,55 +1,0 @@
-Position summary
-Do you want to build your brand by working for a leading consulting firm that drives eminence in the marketplace? Are you interested in leveraging your analytical skills and strategic ideas to improve mission execution? If so, Deloitte could be the place for you! Our Federal Strategy and Operations teams bring deep industry expertise, rigorous analytical capabilities and a pragmatic mindset to help solve our client’s most complex business problems. Join our team, and play a key role in helping to design our clients’ roadmap to the future and help transform the Federal marketplace.
- 
-Work you’ll do
- 
-As a Sr. Consultant within our Business Model Transformation team, you will:
-Perform project tasks independently, and may direct the efforts of others
-Participate in and/or lead the development of deliverable content that meets the needs of the client and contract
-Anticipate client needs and formulate solutions to client issues
-Review deliverables for accuracy and quality
-Provides coaching to junior staff 
-Contribute to new business proposals and proposal development
-Manages own personal and professional development; seeks opportunities for professional growth and expansion of consulting skills and experiences
- 
- 
-The Team
- 
-Transparency, innovation, collaboration, sustainability: these are the hallmark issues shaping Federal government initiatives today. Deloitte’s Federal practice is passionate about making an impact with lasting change.  Carrying out missions in the Federal practice requires fresh thinking and a creative approach. We collaborate with teams from across our organization in order to bring the full breadth of Deloitte, its commercial and public sector expertise, to best support our clients.  Our aspiration is to be the premier integrated solutions provider in helping to transform the Federal marketplace.
-By offering client-tailored services and deep industry insight, the Business Model Transformation team helps clients to seize new growth opportunities, maximize operational effectiveness, elevate their quality of service delivery, and stay ahead of customer demand. Consider a career where you’ll thrive on helping Federal clients to execute on their strategic visions and improve mission performance.
- 
-Qualifications
- 
-Required:
-Bachelor's degree in Computer Science, Engineering, Mathematics or other business-related field
-4+ years of relevant consulting or industry experience
-Experience working independently with minimal guidance
-Strong problem solving and troubleshooting skills with experience exercising mature judgment
-Proven experience effectively prioritizing workload to meet deadlines and work objectives
-Demonstrated ability to write clearly, succinctly, and in a manner that appeals to a wide audience
-Proficiency in word processing, spreadsheet, and presentation creation tools, as well as Internet research tools
-Ability to obtain and maintain the required TS/SCI Security Clearance for this role
- 
- 
-Preferred:
-Prior professional services or federal consulting experience
-Experience with: SQL; Python; R; VBA, Hadoop
-Familiarity with electronic warfare, TECHELINT or OPELINT also desirable
-Vis tools: Tableau, Qlikview, Spotfire; Cognos 8, BI Authoring and modeling; Cognow CRN
-STATS package such as SPSS (modeler, statistics, text analytics); STATA or SAS; MATLAB
-Gathering data from both information systems and personal interactions
- 
- 
-How you’ll grow
-At Deloitte, our professional development plan focuses on helping people at every level of their career to identify and use their strengths to do their best work every day. From entry-level employees to senior leaders, we believe there’s always room to learn. We offer opportunities to help sharpen skills in addition to hands-on experience in the global, fast-changing business world.  From on-the-job learning experiences to formal development programs at Deloitte University, our professionals have a variety of opportunities to continue to grow throughout their career.  Explore Deloitte University, The Leadership Center.
- 
- 
-Benefits
-At Deloitte, we know that great people make a great organization. We value our people and offer employees a broad range of benefits. Learn more about what working at Deloitte can mean for you.
- 
- 
-Deloitte’s culture
-Our positive and supportive culture encourages our people to do their best work every day. We celebrate individuals by recognizing their uniqueness and offering them the flexibility to make daily choices that can help them to be healthy, centered, confident, and aware. We offer well-being programs and are continuously looking for new ways to maintain a culture where our people excel and lead healthy, happy lives.  Learn more about Life at Deloitte.
- 
-Corporate citizenship
-Deloitte is led by a purpose: to make an impact that matters. This purpose defines who we are and extends to relationships with our clients, our people and our communities.  We believe that business has the power to inspire and transform. We focus on education, giving, skill-based volunteerism, and leadership to help drive positive social impact in our communities.  Learn more about Deloitte’s impact on the world.
