+++ requirements.txt
@@ -1,9 +1,30 @@
-numpy==1.13.1
-Word2Vec==0.9.1
+Cython==0.26
+Flask==0.12.2
+Flask-Cors==3.0.3
+Jinja2==2.9.6
+MarkupSafe==1.0
 PorterStemmer==0.5
-stop_words==2015.2.23.1
+SQLAlchemy==1.1.12
+Werkzeug==0.12.2
+chardet==2.2.1
+click==6.7
+colorama==0.2.5
+command-not-found==0.3
+gensim==2.3.0
+html5lib==0.999
+itsdangerous==0.24
+language-selector==0.1
 nltk==3.2.4
-sqlalchemy==1.1.11
-flask==0.12.2
-scipy==0.19.1
-gensim==2.2.0
+numpy==1.13.1
+pycurl==7.19.3
+pygobject==3.12.0
+python-apt===0.9.3.5ubuntu2
+requests==2.2.1
+six==1.5.2
+stop-words==2015.2.23.1
+uWSGI==2.0.15
+ufw===0.34-rc-0ubuntu2
+unattended-upgrades==0.1
+urllib3==1.7.1
+wheel==0.24.0
+word2vec==0.9.1
